Improvement: Multi Param Model Service Map (OFBIZ-12761)
When you define a service with a map as parameter or a List like the following example :

    <service name="exampleService" location="component:..." invoke="exampleService">
        <description>Service example use case with internal attributes to validate </description>
        <attribute name="partyId" type="String" mode="IN"/>
        <attribute name="header" type="java.util.Map" mode="IN"/>
    </service>

You can't define the "header" map structure to ensure data quality.

The services.xsd currently supports attribute elements as children so we can improve the previous example

    <service name="exampleService" location="component:..."
        invoke="exampleService" transaction-timeout="7200">
        <description> Service example use case with internal attributes to validate </description>
        <implements service="ApiJsonInterface"/>
        <attribute name="partyId" type="String" mode="IN"/>
        <attribute name="header" type="java.util.Map" mode="IN">
            <attribute name="internalCode" type="String" optional="false"/>
        </attribute>
    </service>

This commit improves the service engine to analyze the Map content or List of Map, if the service defines is

        <attribute name="header" type="java.util.Map" mode="IN">
            <attribute name="internalCode" type="String" optional="false"/>
        </attribute>

It accepts a context like `[partyId: "foo", [header: [internalCode: "bar"]]]`
or

        <attribute name="header" type="java.util.List" mode="IN">
            <attribute name="internalCode" type="String" optional="false"/>
        </attribute>

It accepts a context like `[partyId: "foo", header: [[internalCode: "bar"]]]`

By the way while improving the code to support this, we took the opportunity to refactor outdated code using java streams and cover it by unit tests.

Thanks to Gaetan Chaboussie for helping me with this improvement