Fixed: Convert CommunicationEventServices mini lang to groovy (OFBIZ-9992) (OFBIZ-11164)
In the new groovy file, CommunicationEventServices.groovy The parameter was set
as the whole object communicationEvent instead of the ID only
communicationEventId which is the one required. It gave this error :
Error running Groovy method [sendEmailDated] in Groovy file
[component://party/groovyScripts/communication/CommunicationEventServices.groovy]:
(org.apache.ofbiz.service.ServiceValidationException: The following required
parameter is missing: [IN] [sendCommEventAsEmail.communicationEventId])

I did the line change and it worked and sent the email.

jleroux: I was a bit lost in (OFBIZ-9992) and (OFBIZ-11164) so I simply checked
sendCommEventAsEmail definition and that's alright

Thanks: Serge Byishimo (https://github.com/sergeByishimo)