Fixed: Added accounting main action menu to Organization GL Settings screens (OFBIZ-12771)
Clicking on the Create a Gl Transactions button navigates to a screen
within the Organisation GL Settings group of screens. Within this
context the main actions menu - which included the Create button - is
not visible, meaning it is cumbersome for the user to add another
transaction.

This commit adds the main actions menu to the Organization GL Settings
screens, matching the way the actions are displayed in all other
Accountancy screens.