Fixed: Cannot add a new geo point to a facility (OFBIZ-12793)
In demo-trunk, visit https://demo-trunk.ofbiz.apache.org/facility, select the
'My Retail Store' facility and then navigate to that facility's Geo Location tab.

Adding a new geo location to the facility fails with the error:
    No such property: ServiceUtil for class: FacilityGeoServices

Error is due to a missing import of the ServiceUtil class in
FacilityGeoServices.groovy.

Thanks: Daniel for report and 1st analysis. There was also a use of result
instead of serviceResult