Fixed: Eclipse build problems and proper dependency setup (OFBIZ-12808)
Due to improper dependency configurations and the JPMS (Java Plattform
Module System) which was introduced to Java since version 9, the Eclipse
build and running/debugging is not working with JDK 17 (trunk and
release22.01).

The reason is that there are dependencies to libraries which are also
shipped with the JDK which causes a conflict leading to ignore those
packages/classes in the build.

This commit fixes the problems for release22.01 without plugins, they
have to be modified separately.