Improve: Resolve uiLabels through ajax with dedicated screen
Before this improvement, you can resolve a list of labels to display with the uiLabel resource file
  like : "CommonUiLabels" : ["CommonUpload", "CommonSave", "CommonCompleted"]

We remove the possibility to specify a uiLabel resource file and make it work by default with CommonUiLabels.
After that we implement the json return on extend theme screen so we can add some uiLabel ressource by theme extension or component screen definition.

Other enhancement, on getJSONuiLabels instead resolve the label with a list position you can call your label with it key.
      resolve : uiLabelJsonObjects.CommonUiLabels[1]
       become : uiLabelJsonObjects.CommonSave