Implemented: Replace SvnCheckout in Gradle (OFBIZ-12868) (INFRA-25327)
The pullAllPluginsSource.script is used in several CI places (GitHub actions,
BuildBot and Demos)

Lastly I committed the pullAllPluginsSource.sh and pullPluginSource.sh scripts
with Windows. As Windows has not the x flag, they were not set as executable.

This fixes it by setting them as executable using "git update-index --chmod=+x"

This also changes docker-image.yaml and gradle.yaml to use pullAllPluginsSource

I removed the build in docker-image.yaml as it's not longer needed. Only
pullAllPluginsSource was used and this task does not build anything. It just
pulls the sources

I did not set the the x flag for Windows  scripts as it will be useless anyway
since Windows does not use the x flag

Conflicts handled by in .github/workflows/gradle.yaml