Fixed: Replace SvnCheckout in Gradle (OFBIZ-12868)
Updates the documentation for pullAllPluginsSource.sh and pullPluginSource.sh

Uses
git switch -C "$branch"
instead of
git switch "$branch"
in both scripts

There is something weird with pullPluginSource.sh. It works perfectly on Windows
using git-bash, but in Ubuntu 20.04 it does not add but replaces plugins.
I tested using current last Git version in both OS
(respectively 2.43.0.windows.1 and 2.43.0). It's a minor issue as it does not
block the CI so I'll review that later...