Improved: Screen Communications in Party should not show create trigger to user with VIEW permissions (OFBIZ-12891)
When accessing https://demo-trunk.ofbiz.apache.org/partymgr/control/FindCommunicationEvents as a user with only
VIEW permissions (.e.g. userId = auditor) the action trigger to create a new communication is shown.

This should not be visible to such a user as it leads to an undesired effect and diminished user experience.

modified:
- PartyMenus.xml - added permission condition to menu-item 'new' in CommSubTabBar