Improved: Swap dateTime field in French (OFBIZ-12822 - #640)
This improvement focuses on enhancing the syntax of Apache OFBiz by swapping the order of the dropdown field and the text field for French date time input. Specifically, this change applies only to the French language.

By default, the date time input in Apache OFBiz consists of a text field followed by a dropdown field for selecting the time. However, for French users, it is more intuitive to have the dropdown field for selecting the time appear before the text field.

This improvement modifies the syntax to accommodate the French language, ensuring a more user-friendly and natural experience for French-speaking users.

Please note that this change is language-specific and will not affect other languages or locales.

This improvement aims to align the date time input order with French user expectations and contribute to a smoother user experience in Apache OFBiz.