Improved: FinAccount Screen shows create trigger to user with VIEW permission (OFBIZ-12567)
When accessing https://localhost:8443/accounting/control/EditFinAccount?finAccountId=ABN_CHECKING as a user with only VIEW permissions (e.g. userid=auditor) the screen shows a action trigger to create something.
This should not be visible to such a user as it leads to a undesired effect and diminished user experience.

modified:
- FinAccountScreens.xml - removed container having link target to create a new financial account
- AccountMenus.xml - added menu-iitem to MainActionMenu for creating a new financial account