Improved: Update build.gradle to the latest dependencies (OFBIZ-12921)
I did the last update with OFBIZ-12658 in June 2022.
IIRW I did not speak about the reason I did not update since.

It's simple. We encounter more and more conflicts when upgrading.
Often it's hard work to resolve these conflicts for a disputable benefit.

After all, it's working why upgrading?
As the famous meme "if it ain’t broke, don’t fix it" says.

So I decided to only upgrade dependencies when security issues make it necessary.
Few weeks ago I finally decided to though have a look at the situation.

Clearly it confirms what I thought. If it's not a trivial upgrade that does not
bring much, most of the time it's conflicts.

So I'll now rather upgrade only in case of security necessity.
Of course you are welcome to continue to upgrade as much as possible.
Maybe sometimes issues crossed with previous versions will resolved with newer
libraries, but that must be very rare...if ever...