Improved: MFG-Move from hard-code menu location to parameterized (OFBIZ-12928)
Customization is a key aspect of the OFBiz (Open For Business) platform, providing businesses with a robust and adaptable base to customize their enterprise software solutions according to their unique operational requirements.
The *Menus.xml file of a component, essential for designing user interfaces that are both intuitive and tailored to the specific roles and tasks of its users, plays a significant role in this context, as referenced within screen widgets and Freemarker templates.
However, as customizations and the project's source code evolve over time, the likelihood of encountering merge conflicts increases significantly, especially when attempting to incorporate customizations alongside ongoing bug fixes and enhancements from the project.
To improve the appeal of OFbiz and the developer experience, it is recommended to shift from fixed menu locations to a parameterized approach in our components and plugins. This change would enable developers to more seamlessly incorporate a 'custom' menu, facilitating the integration of project updates.

modified:
manufacturing/WEB-INF/web.xml: added parameter for mainMenuLocation
various *Screens.xml files: changing location of 'include-menu' elements from location="component://manufacturing/widget/ManufacturingMenus.xml to location="${parameters.mainMenuLocation}"