Fixed: ‎SecuredUpload‎::checkMaxLinesLength put in a regression for PDF files (OFBIZ-12936)
As commented at https://s.apache.org/wpkuj, while working on OFBIZ-12926
I realised that checkMaxLinesLength put in a possible regression for PDF files.

As a summary: we can't use checkMaxLinesLength for PDF files because they may
contain several CharSet encodings. Hence no possibility to use
Files::readAllLines that needs a sole CharSet. So we need to bypass this check.