Fixed: BlueLight theme error when launching OFBiz (OFBIZ-12938)
We have an error for BlueLight theme.
HumanResShortcutAppBar and AccountingShortcutAppBar are concerned.
Tomahawk (derived from BlueLight) is OK.

This simply replace the corresponding lines in BlueLight by the sole one in
Tomahawk.  I also noted that there is in BlueLight the same line at line 84
before patching

Thanks: Pierre Smits for help