improved: have profileMenuLocation (OFBIZ-12981) (#749)
* Improved: ORDER - have requestMenuLocation (OFBIZ-12991)

Given that request screens are used as portal pages, the inclusion of such portal pages in other applications leads to errors being thrown for the associated menus (due to the use of the mainMenuLocation parameter in web.xml).
Having a requestMenuLocation in web.xml solves this issue.

modified:
- web.xml: add contextParam requestMenuLocation
- CommonScreens.xml: change location request related menus
- CustRequestScreens.xml: change location request related menu

* Improved: have profileMenuLocation (OFBIZ- 12981)

Profile screens are used in various components as portalpageportlets, and have references to various party menus. As menus in the various profile screens are parameterized to use ${parameters.mainMenuLocation} this generates errors in the components/applications where these screens are used as portal page elements.
Having a context-param profileMenuLocation resolves this issue.

modified:
web.xml: added context-param profileMenuLocation

* Improved: have profileMenuLocation (OFBIZ- 12981)

Profile screens are used in various components as portalpageportlets, and have references to various party menus. As menus in the various profile screens are parameterized to use ${parameters.mainMenuLocation} this generates errors in the components/applications where these screens are used as portal page elements.
Having a context-param profileMenuLocation resolves this issue.

modified:
ProfileScreens.xml: change menu locations