Improved: Remove Agreements menu-item from AR and AP menus (OFBIZ-13010)
The Agreements menu-item serves no specific purpose in both the AR and the AP application of the accounting component. As a user I would expect that in the AR application it would present the search/find functionality related to contracts with customers and the same in the AP application but then related to contracts with suppliers.
However, this is not the case. The functionality just presents the OOTB functionality of the accounting application without any uniqueness related to resp AR or AP.

modified:
- ApMenus.xml - remove menu-item agreements
- ArMenus.xml - remove menu-item agreements