Improved: Reduce logo size BlueLight theme (OFBIZ-12997)
When using the Bluelight theme the size of the default logo is greater than the visible 'system notes' lines.
Reducing the height of the logo displayed, increases the visibility of the pages useful to user and thus enhance the user experience (less vertical scrolling).

modified: header.ftl - added height value