Fixed: invoiceOverview generates an error (OFBIZ-13017)
Fixes the problem by replacing replacing all invoiceOverview occurrences in
OFBiz by ViewInvoice. There are no longer other occurrences of "invoiceOverview"

This is a maybe temporary workaround. Possibly more is needed as we want to use
AccountingInvoiceCpd.xml and related changes that were introduced with
OFBIZ-12994. I suppose it's enough for the moment since it does not introduce
other changes.

As she proposed me, I have asked Leïla to have a look because she knows that
more than me having worked on OFBIZ-12994. Maybe a complete refactor is needed.