Fixed: Old error in showPortalPage screen (OFBIZ-13059)
I spotted this error in demo stable and trunk error.log. It's due to an old
commit: https://svn.apache.org/repos/asf/ofbiz/trunk@1023286

I tried a bit to understand what's the problem was but as it's a 13,5 years old
bug I eventually decided to simply remove the wrong parts. I'll not backport.