Fixed: Issue when caching a query with a selectField (OFBIZ-13072)
When a query with a field selection and a cache() option is done, the
result is put in cache with the incomplete genericValue (I mean without
all the fields).

Then when an other query with the same condition and cache option is
done, even whitout the field selection, the result returned is the
incomplete version of the GenericValue that has been stored in cache.

This prevent from adding in cache an incomplete genericValue and filter
the cached value when necessary