Fixed: Tomcat access_log has 2 dots (OFBIZ-13089)
Between Tomcat 7 and Tomcat 8 the access_log default value of fileDateFormat
has changed:
https://tomcat.apache.org/tomcat-7.0-doc/config/valve.html#Access_Log_Valve/Attributes
https://tomcat.apache.org/tomcat-8.0-doc/config/valve.html#Access_Log_Valve/Attributes

I did not find any notice about that. We did not change OFBiz access-log-prefix value.
Not a big deal but better keep a simple file format with only one dot.

As we use the defaul format value we could also remove this OOTB value.
I believe it's better to keep it in order to allow users to decide about it.
Without this OOTB value they may miss this opportunity.