Improved: Inform user about no agreements when creating an order (OFBIZ-13116)
On 2nd create order screen, despite the title
<<Enter Order Currency, Agreements, and Ship Dates>>
currently when no agreement exist this information is not shown.
That's a pre Apache situation. It can be useful to be known and it's an easy
and safe change: only static information so no possible regression.

This was respectively reminded to us by these messages on dev and user MLs
https://lists.apache.org/thread/9cz2q5kcovnqm86lfj2nloozvzswgmpm
https://lists.apache.org/thread/5ll9o8onhzlll4h9vxngnz31ytjrl67f

Thanks: Bill Harder for report