Improved: Added support to allow direct view rendering in override view functionality (OFBIZ-13117)
Added allow-direct-view-rendering and direct-view-rendering-with-auth in
view-mapping tag, default values will be false. i.e by default now view
is allowed to be used as OOTB overridden view functionality.
In order to allow the view redirection (override) on all workflows
allow-direct-view-rendering must be set to true.
If view redirection is allowed and direct-view-rendering-with-auth is
set to true then login credentials are necessary to use this
functionality.
This feature may break some existing flow where overridden view workflow
is used