Fixed: Fix some bugs SpotBugs reports (OFBIZ-12386)
After updating SpotBugs plugin to 4.8.6.r202406180231-6cf7b2c
in Eclipse 2023-03 (4.27.0) it reported:

<<Bug: Class (org.apache.ofbiz.webapp.WebAppCache) using singleton design pattern
has non-private constructor.

This class is using singleton design pattern and has non-private constructor
(please note that a default constructor might exist which is not private).
Given that, it is possible to create a copy of the object, thus violating the
singleton pattern.  The easier solution would be making the constructor private.
>>

As it's also used by WebAppCacheTest class I made it package. That's safer but
not enough so I also put it in exclude.xml
Also Chekstyle wants it to be final