Fixed: Trunk demo errors in error.log (OFBIZ-13005)
With my last commit b365f9db8988bc5ac4d1d6db7bfd53ee32b7d612 I wrote that with
<<    keywords.add(contentWrapper.get('PRODUCT_NAME', 'html'))
keywords is never empty.>>

I found again the same error in trunk demo so we need to also check that
contentWrapper contains 'html' for 'PRODUCT_NAME'. Here it is.

All this comes from when moving from Minilang to Groovy, which proves that it
was not an easy task.