Fixed: Upload image size issue (OFBIZ-12639)
SecuredUpload::checkMaxLinesLength does not work when the charset used to create
the file is not the same than the one used when uploading. It's a know problem.

This at least allow images to be uploaded.

I'll check if we can improve the call in SecuredUpload::checkMaxLinesLength to
FileUtils.readLines() (Apachecommons.io) according to
http://illegalargumentexception.blogspot.com/2009/05/java-rough-guide-to-character-encoding.html#javaencoding_autodetect

See https://lists.apache.org/thread/dv4yjpknms5zd2l73wb8ht3s0db2wx2v for details