Fixed: When you add a party content the party profile screen does not refresh (OFBIZ-6709)
There is a problem showing the result of upload either if it's rejected or not.
In all case it shows "Maybe for security reason your file has not been accepted,
 check the log."

The best solution is to temporarily show this information, say 10 sec.
No need to show if it's a success, the list above shows it anyway.

That's what I did but  7 sec is enough to read the message.
Massage that is a bit longer now, but more explicit:
"If you don't see your file in Party Content list above, it has been rejected
for security reason. Check the log."