Improved: Add display result format for auto-completion on lookup
Currently, when you have a lookup with autoCompletion, the result contains the fields to display or by default search fields.

The format is strict. This improvement add new field  displayFieldFormat to send a UEL schema that we want.

This commit introduce an example with LookupWorkEffort :

****
      <set field="displayFieldFormat" value="\${workEffortName} - \${groovy: description?.length() &gt; 20 ? description.substring(0, 20) + '…': description}"/>
****

We want display the workEffortName and the description limited to 20 chars. To forward the UEL string, we need to use the escape char \ like this the screen engine doesn't parse displayFieldFormat and let the ftl AjaxAutocompleteOptions.ftl realized the translation.