Fixed: Remove duplicate event handler for collapsible in modals
OF-13169

When a modal opens, we search interactive elements
and trigger jQuery plugins/behavior in bindObservers
function. This function searches for HTML elements and
starts needed behavior, using this logic:

$(wrappingElement).find(selector).each([behavior]);

This means that all 'selector' in 'wrappingElement'
initially present in the page will be triggered.
When we open a modal, we call bindObservers again
within the modal, to trigger interactive elements
which are in the modal.

For collapsible (and some others behaviors), we use

$(wrappingElement).on('click', doStuff);

...but the jQuery 'on' method also affects
DOM elements that are dynamically added to the
page, like modals content. So for these behaviors,
event handlers for interactive elements are created
twice. So when we click on the "+" of a collapsible
in a modal, "toggleCollapsiblePanel" is called twice,
and the collapsible section is open and closed right
away