Fixed: jsTree visual glitch on link hover
OF-13174

Helveticus introduced a global CSS transition effect,
which caused various problems across the app (all the
elements that changed dimensions were animated).
jsTree plugins is concerned, when hovering links in
a tree.

This PR removes the global transition effect, and put
it on identified elements that need it (buttons and links).