Improved: Add delegator method to create in mass many entities (OFBIZ-13176)
When you need to load huge record on your database (like system synchronization, data import) OFBiz can have latency with the delegator.create.
Incriminated unitary workflow and eeca control. You can disable eeca on use a dedicate delegator but sometime it's not enough.

So for special case, we implement a new function delegator.createAllByBatchProcess that will use an insert with one transaction through the standard sql method INSERT TO ... VALUES ... .

By the way, like the process isn't unitary, the eeca system can't run on this batch, so this function must therefore be used with care

Thanks to Néréide team for this contribution