Improved: Add a date range picker form field
This is a wrapper around https://www.daterangepicker.com/, exposing
(almost) all its parameters as field attributes.

It can be used as a date picker (like <date-time/>), or as a date range
picker (like <date-find/>).

Its main purpose is to provide a modern UX to pick date ranges, with
2 calendars where the user can click start date, then end date, instead
of dealing with two unrelated date fields and operator selectors.

OFBIZ-13209