Improved: Rewrite UtilCacheTest (OFBIZ-13311)
Rewrites the test class UtilCacheTest.
Changes the variable names to make them more clear and moves the
test utility objects to a dedicated new class.
Also removes the GenericTestCaseBase, that was used only in
the UtilCacheTest.
Deleted a part of the testExpire test method, that was based on
the useAllMemory method, which is highly OS or hardware dependant
making unit test dependant as well.

Thanks: Nicolas for the help in analyzing.