Backport to release24.09 commit 5c66eaa681ad0ac7fec5410b764463e651cac300:
Improved: Refactor readHtmlDocument in UelFunctions to obtain an instance of org.w3c.dom.Document using the standard javax.xml.parsers.* APIs instead of org.cyberneko.html.

Removing the Cyberneko dependency allows us to drop the Maven repository
https://repository.ow2.org/nexus/content/repositories/public/, which is
offline at the time of this commit (due to a license issue) and was
causing OFBiz builds to fail. This repository has also had issues in the
recent past, as reported in
https://issues.apache.org/jira/browse/OFBIZ-13206.